requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
graphviz==0.20.1
sqlalchemy==2.0.27
pylint==3.0.3
